+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## v4.0.2
+## v4.1.0
 
 ### BUG FIXES